Added function for interactive login
Using current functionality, logging in requires a user to manually
enter their credentials within their script (using the login() function or
by passing credentials to the Panoptes object initializer), or to store them as
environment variables in `PANOPTES_USERNAME` and `PANOPTES_PASSWORD`.

This commit adds functionality for an interactive login. This would allow a
user to enter their credentials manually when running a script that doesn't
involve storing the credentials in a static file or in an environment
variable.