Fix: Can't save new objects with individual links
Fixes #205

If the link slug wasn't already in the list of links, trying to set the
link would set an attribute on the `PanoptesObject` instead of creating an
entry in `raw['links']`.

This PR changes the setattr logic to make sure the slug _isn't_ an
attribute rather than checking it _is_ a link. This allows setting links
that the client doesn't already know about, but has the side effect of
allowing invalid links. This is actually a good thing, because invalid
links will be rejected by the API, whereas before they would have been
set as attributes and silently discarded on save.