Deep copy of initial subject metadata
Determining whether to save the subject metadata relies on comparison between a potentially updated metadata dict and a copy of the original taken when the Subject instance was created. Using assignment or even copy construction can (apparently not always) result in the "copy" in fact being a reference to the object which may be modified. This means that the two always compare equal and the metadata is never saved. Using copy.deepcopy solves this issue.